Fix a bug/ cleanup

+++ WeatherApp.Models/MainInfo.cs
@@ -9,8 +9,8 @@ namespace WeatherApp.Models
     public class MainInfo
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
